cast qps to float32

+++ cluster-autoscaler/cloudprovider/clusterapi/clusterapi_provider.go
@@ -164,14 +164,14 @@ func BuildClusterAPI(opts config.AutoscalingOptions, do cloudprovider.NodeGroupD
 	if err != nil {
 		klog.Fatalf("could not generate dynamic client for config")
 	}
-	managementConfig.QPS = opts.KubeClientQPS
+	managementConfig.QPS = float32(opts.KubeClientQPS)
 	managementConfig.Burst = opts.KubeClientBurst
 
 	workloadClient, err := kubernetes.NewForConfig(workloadConfig)
 	if err != nil {
 		klog.Fatalf("create kube clientset failed: %v", err)
 	}
-	workloadConfig.QPS = opts.KubeClientQPS
+	workloadConfig.QPS = float32(opts.KubeClientQPS)
 	workloadConfig.Burst = opts.KubeClientBurst
 
 	managementDiscoveryClient, err := discovery.NewDiscoveryClientForConfig(managementConfig)
